modified:   src/App.jsx deleted:    src/components/About.jsx new file:   src/components/ProgramInfo.jsx deleted:    src/components/WhatsNew.jsx
